+++ src/app/components/landing-page/landing-page.component.html
@@ -5,7 +5,7 @@
     <p>
       Ein Anmeldesystem für jeden und jede Angelegenheit. Mit diesem System kannst du ohne Probleme deine Termine
       Verwalten und somit immer die
-      Übersicht behalten. Egal ob Einfaches eintragen in eine Liste, modulare individuelle Ja/Nein Zusatzoptionen
+      Übersicht behalten. Egal ob einfaches Eintragen in eine Liste, modulare individuelle Ja/Nein Zusatzoptionen
       oder sonstiges. Der Termin
       ist sehr schnell und unkompliziert erstellt und man kann sich sofort dafür anmelden.
     </p>
